chore: move update to new update files and introduce more forward declarations
