+++ libs/model/include/bulin/model/cereal.hpp
@@ -1,17 +1,17 @@
 #pragma once
 
-#include <Magnum/Math/RectangularMatrix.h>
 #include <Magnum/Math/Vector.h>
-
 #include <cereal/cereal.hpp>
+
+#include <bulin/model/model_fwd.hpp>
+
+#include <filesystem>
 
 // This code has mostly been adapted from <cereal/types/vector.hpp>
 // We don't deal for now with data that could be potentially serialized
 // directly in binary format.
-
 namespace cereal
 {
-
 template<typename Archive, std::size_t Size, typename T>
 void save(Archive& ar, Magnum::Math::Vector<Size, T> const& vector)
 {
@@ -28,5 +28,10 @@ void load(Archive& ar, Magnum::Math::Vector<Size, T>& vector)
     vector[i] = std::move(x);
   }
 }
-
 }  // namespace cereal
+
+namespace bulin
+{
+void save(std::filesystem::path const& fname, model state);
+auto load(std::filesystem::path const& fname) -> model;
+}
